no_std ergo-lib, make ergo-merkle-tree and ergo-nipopow optional
