+++ ergo-lib/Cargo.toml
@@ -17,9 +17,9 @@ sigma-util = { workspace = true }
 ergo-chain-types = { workspace = true }
 ergotree-ir = { workspace = true }
 ergotree-interpreter = { workspace = true }
-ergo-nipopow = { workspace = true }
+ergo-nipopow = { workspace = true, optional = true }
 ergoscript-compiler = { workspace = true, optional = true }
-ergo-merkle-tree = { workspace = true }
+ergo-merkle-tree = { workspace = true, optional = true }
 ergo-rest = { workspace = true, optional = true }
 indexmap = { workspace = true }
 base16 = { workspace = true }
@@ -28,39 +28,44 @@ serde_json = { workspace = true, optional = true }
 thiserror = { workspace = true }
 derive_more = { workspace = true }
 bounded-vec = { workspace = true }
-num-bigint = { workspace = true, features = ["serde"] }
 proptest-derive = { workspace = true, optional = true }
 k256 = { workspace = true }
 sha2 = { workspace = true }
 hmac = { version = "0.12" }
 pbkdf2 = "0.11"
-rand = { workspace = true }
+rand = { workspace = true, optional = true }
 bitvec = { workspace = true, optional = true }
-unicode-normalization = "0.1.19"
+unicode-normalization = { version = "0.1.19", default-features = false }
 lazy_static = { workspace = true }
 proptest = { workspace = true, optional = true }
 serde_with = { workspace = true, optional = true }
-itertools = { workspace = true }
+hashbrown = { workspace = true }
 
 
 [features]
-default = ["json"]
+default = ["std", "json", "nipopow", "merkle"]
+std = ["rand", "ergotree-ir/std", "ergotree-interpreter/std"]
 json = [
     "serde",
     "serde_json",
     "serde_with",
     "bounded-vec/serde",
     "ergotree-ir/json",
+    "ergotree-interpreter/json",
+    "ergo-merkle-tree?/json",
 ]
 compiler = ["ergoscript-compiler"]
 arbitrary = [
+    "std",
     "proptest",
     "proptest-derive",
     "ergotree-ir/arbitrary",
     "ergo-chain-types/arbitrary",
     "ergotree-interpreter/arbitrary",
 ]
-mnemonic_gen = ["bitvec"]
+merkle = ["ergo-merkle-tree"]
+nipopow = ["ergo-nipopow"]
+mnemonic_gen = ["bitvec", "rand"]
 rest = ["ergo-rest"]
 
 [dev-dependencies]
@@ -73,3 +78,11 @@ pretty_assertions = { workspace = true }
 bs58 = { workspace = true }
 byteorder = { workspace = true }
 expect-test = { workspace = true }
+
+
+# docs.rs-specific configuration
+[package.metadata.docs.rs]
+# document all features
+all-features = true
+# enable docsrs flag, which lets us enable doc_auto_cfg to mark feature-gated functionality in documentation
+rustdoc-args = ["--cfg", "docsrs"]
