+++ ergo-lib/src/chain/contract.rs
@@ -37,16 +37,13 @@ impl Contract {
     }
 }
 
-#[cfg(test)]
+#[cfg(all(test, feature = "compiler"))]
 #[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
 
-    #[cfg(feature = "compiler")]
     #[test]
     fn compile() {
-        let contract =
-            Contract::compile("HEIGHT", ergoscript_compiler::script_env::ScriptEnv::new()).unwrap();
-        dbg!(&contract);
+        Contract::compile("HEIGHT", ergoscript_compiler::script_env::ScriptEnv::new()).unwrap();
     }
 }
