+++ ergo-lib/src/chain/json/hint.rs
@@ -1,9 +1,10 @@
-use std::collections::HashMap;
+use alloc::vec::Vec;
+use hashbrown::HashMap;
 
 use ergotree_interpreter::sigma_protocol::prover::hint::{Hint, HintsBag};
 use serde::{Deserialize, Serialize};
 
-use crate::wallet::multi_sig::TransactionHintsBag;
+use crate::wallet::TransactionHintsBag;
 
 #[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
 pub struct TransactionHintsBagJson {
@@ -51,7 +52,7 @@ impl From<TransactionHintsBagJson> for TransactionHintsBag {
 
 #[cfg(test)]
 mod tests {
-    use crate::wallet::multi_sig::TransactionHintsBag;
+    use crate::wallet::TransactionHintsBag;
     use proptest::prelude::*;
 
     proptest! {
