+++ ergo-lib/src/chain/json/transaction.rs
@@ -1,4 +1,6 @@
-use std::convert::TryFrom;
+use alloc::string::String;
+use alloc::vec::Vec;
+use core::convert::TryFrom;
 use thiserror::Error;
 
 use crate::chain::transaction::unsigned::UnsignedTransaction;
