+++ ergo-lib/src/wallet/derivation_path.rs
@@ -2,8 +2,14 @@
 //! BIP-44 <https://github.com/bitcoin/bips/blob/master/bip-0044.mediawiki>
 //! and EIP-3 <https://github.com/ergoplatform/eips/blob/master/eip-0003.md>
 
+use alloc::{
+    boxed::Box,
+    collections::VecDeque,
+    string::{String, ToString},
+    vec::Vec,
+};
+use core::{fmt, num::ParseIntError, str::FromStr};
 use derive_more::From;
-use std::{collections::VecDeque, fmt, num::ParseIntError, str::FromStr};
 use thiserror::Error;
 
 /// Index for hardened derivation
