+++ ergo-lib/src/wallet/secret_key.rs
@@ -1,5 +1,6 @@
 //! Secret types
 
+use alloc::vec::Vec;
 use derive_more::From;
 use ergo_chain_types::EcPoint;
 use ergotree_interpreter::sigma_protocol::private_input::DhTupleProverInput;
@@ -29,11 +30,13 @@ pub enum SecretKey {
 
 impl SecretKey {
     /// Generates random DlogProverInput
+    #[cfg(feature = "std")]
     pub fn random_dlog() -> SecretKey {
         SecretKey::DlogSecretKey(DlogProverInput::random())
     }
 
     /// Generates random DhTupleProverInput
+    #[cfg(feature = "std")]
     pub fn random_dht() -> SecretKey {
         SecretKey::DhtSecretKey(DhTupleProverInput::random())
     }
@@ -136,7 +139,7 @@ impl From<SecretKey> for PrivateInput {
 #[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
-    use std::convert::TryInto;
+    use core::convert::TryInto;
 
     #[test]
     fn dlog_roundtrip() {
