+++ ergo-merkle-tree/Cargo.toml
@@ -10,19 +10,19 @@ description = "Merkle tree proofs"
 
 [dependencies]
 blake2 = { workspace = true }
-base16 = { workspace = true, optional = true }
+base16 = { workspace = true, features = ["std"], optional = true }
 serde = { workspace = true, optional = true }
 serde_repr = { version = "0.1.7", optional = true }
 serde_json = { workspace = true, optional = true }
-# currently thiserror is only needed for json conversion, so it's feature-gated behind json. 
+# currently thiserror is only needed for json conversion, so it's feature-gated behind json.
 # This may change in the future
-thiserror = { workspace = true, optional = true } 
+thiserror = { workspace = true, optional = true }
 itertools = { workspace = true }
-proptest-derive = {workspace = true, optional = true }
+proptest-derive = { workspace = true, optional = true }
 sigma-ser = { workspace = true }
 ergo-chain-types = { workspace = true }
 sigma-util = { workspace = true }
-proptest = { workspace = true , optional = true }
+proptest = { workspace = true, optional = true }
 
 [features]
 default = ["json"]
