+++ ergotree-interpreter/tests/signing_spec_tests.rs
@@ -1,3 +1,4 @@
+use core::convert::TryInto;
 use ergotree_interpreter::sigma_protocol::private_input::DlogProverInput;
 use ergotree_interpreter::sigma_protocol::verifier::{TestVerifier, Verifier};
 use ergotree_ir::chain::context::Context;
@@ -13,7 +14,6 @@ use ergotree_ir::sigma_protocol::sigma_boolean::{ProveDhTuple, SigmaProp};
 use ergotree_ir::types::stype::SType;
 use num_bigint::BigUint;
 use sigma_test_util::force_any_val;
-use core::convert::TryInto;
 
 #[test]
 fn sig_test_vector_provedlog() {
