+++ ergotree-ir/src/chain/ergo_box/register/value.rs
@@ -1,4 +1,3 @@
-
 use alloc::string::String;
 use alloc::string::ToString;
 use alloc::vec::Vec;
