+++ ergotree-ir/src/mir/coll_append.rs
@@ -1,6 +1,5 @@
 use alloc::boxed::Box;
 
-
 use crate::serialization::op_code::OpCode;
 use crate::serialization::sigma_byte_reader::SigmaByteRead;
 use crate::serialization::sigma_byte_writer::SigmaByteWrite;
