+++ ergotree-ir/src/mir/coll_size.rs
@@ -1,6 +1,5 @@
 use alloc::boxed::Box;
 
-
 use crate::serialization::op_code::OpCode;
 use crate::types::stype::SType;
