+++ ergotree-ir/src/mir/create_avl_tree.rs
@@ -2,7 +2,6 @@
 
 use alloc::boxed::Box;
 
-
 use super::expr::Expr;
 use crate::has_opcode::HasStaticOpCode;
 use crate::mir::expr::InvalidArgumentError;
