+++ ergotree-ir/src/mir/option_is_defined.rs
@@ -1,6 +1,5 @@
 use alloc::boxed::Box;
 
-
 use super::expr::Expr;
 use super::expr::InvalidArgumentError;
 use super::unary_op::OneArgOp;
