+++ ergotree-ir/src/mir/subst_const.rs
@@ -1,7 +1,6 @@
 //! Substitution of constants in serialized sigma expression
 use alloc::boxed::Box;
 
-
 use super::expr::Expr;
 use crate::has_opcode::HasStaticOpCode;
 use crate::mir::expr::InvalidArgumentError;
