+++ ergotree-ir/src/mir/upcast.rs
@@ -2,7 +2,6 @@
 
 use alloc::boxed::Box;
 
-
 use super::expr::Expr;
 use super::expr::InvalidArgumentError;
 use crate::serialization::op_code::OpCode;
