+++ ergotree-ir/src/type_check.rs
@@ -1,6 +1,5 @@
 //! Type checking
 
-
 use alloc::string::String;
 
 use crate::mir::expr::Expr;
